Added Tkinter gui (#1)
This program generates a suggested username based on the user's first name and last name. The program takes up to the first 7 letters of the last name and the first letters of the first name to create a username, keeping the name spelling. If the generated username is not unique, the program removes letters from the last name and adds letters from the first name until a unique username is created.

The program uses the Tkinter module to create a simple graphical user interface. The user inputs their first name and last name into the corresponding entry fields and clicks the "Submit" button. The program then generates a suggested username and displays it to the user. If the user wishes to edit the suggested username, they can click "yes" when prompted and enter a new username.

The program checks for duplicate usernames in a CSV file named "usernames.csv". If a duplicate username is found, the program modifies the username until it is unique. The program also ensures that the generated username only uses letters from the first and last names provided by the user. If any other characters are used, an error message is displayed.

To run the program, simply execute the "tk-uname.py" file. The program requires Python 3 and the Tkinter module to be installed.